requirements.txt
fastapi==0.87.0
nltk==3.7
numpy==1.23.4
PyYAML==6.0
six==1.16.0
uvicorn==0.19.0
